add apitests ass new module
