+++ apitestpetstore/src/main/java/clients/BaseClient.java
@@ -1,4 +1,4 @@
-package apitests.clients;
+package clients;
 
 public class BaseClient {
     static final String BASE_URL = "https://petstore.swagger.io/v2";
