+++ apitestpetstore/src/main/java/model/Category.java
@@ -1,4 +1,4 @@
-package apitests.model;
+package model;
 
 
 import lombok.AccessLevel;
