+++ apitestpetstore/src/main/java/model/Status.java
@@ -1,4 +1,4 @@
-package apitests.model;
+package model;
 
 public enum Status {
     AVAILABLE,
