+++ apitestpetstore/src/main/java/runner/BaseAT.java
@@ -1,4 +1,4 @@
-package apitests.runner;
+package runner;
 
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
