+++ apitestpetstore/src/main/java/runner/pet/PetCrudAT.java
@@ -1,6 +1,6 @@
-package apitests.runner.pet;
+package runner.pet;
 
-import apitests.runner.BaseAT;
+import runner.BaseAT;
 import io.cucumber.junit.CucumberOptions;
 
 @CucumberOptions(features = "src/test/resources/cucumber/features/pet/pet_crud.feature")
