+++ apitestpetstore/src/main/java/steps/pet/PetCrudSteps.java
@@ -1,23 +1,24 @@
-package apitests.steps.pet;
+package steps.pet;
 
-import apitests.clients.PetClient;
-import apitests.model.Pet;
-import apitests.model.ResponseDetails;
-import apitests.steps.BaseSteps;
+import clients.PetClient;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.When;
 import io.restassured.mapper.ObjectMapperType;
+import model.Pet;
+import model.PetUtils;
+import model.ResponseDetails;
+import org.assertj.core.api.Assertions;
+import steps.BaseSteps;
 
 import java.io.File;
 import java.util.Arrays;
 import java.util.List;
 
-import static apitests.model.PetUtils.generatePetWithRandomTestData;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class PetCrudSteps extends BaseSteps {
 
-    private static final Pet newPet = generatePetWithRandomTestData();
+    private static final Pet newPet = PetUtils.generatePetWithRandomTestData();
     private static Pet updatedPet;
 
     public PetCrudSteps(ResponseDetails response, PetClient petClient) {
@@ -32,7 +33,7 @@ public class PetCrudSteps extends BaseSteps {
     @When("Update an existing pet")
     public void update_an_existing_pet() {
         add_a_new_pet_to_the_store();
-        updatedPet = generatePetWithRandomTestData(newPet.getId());
+        updatedPet = PetUtils.generatePetWithRandomTestData(newPet.getId());
 
         response.setResponse(petClient.putPet(updatedPet));
     }
@@ -70,14 +71,14 @@ public class PetCrudSteps extends BaseSteps {
 
         final List<String> statuses = pets.stream().map(Pet::getStatus).toList();
 
-        assertThat(statuses).containsOnly(expectedStatus);
+        Assertions.assertThat(statuses).containsOnly(expectedStatus);
     }
 
     @And("Check the response body of created pet is correct")
     public void check_the_response_body_of_created_pet_is_correct() {
         final Pet createdPet = response.getResponse().getBody().as(Pet.class);
 
-        assertThat(createdPet).isEqualToComparingFieldByFieldRecursively(newPet);
+        Assertions.assertThat(createdPet).isEqualToComparingFieldByFieldRecursively(newPet);
     }
 
     @And("Check the response body of updated pet is correct")
