+++ apitestpetstore/src/test/resources/cucumber/features/pet/test_image.jpg
